% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{predict.segmentRF}
\alias{predict.segmentRF}
\title{Predict a class useing a weighted combination of inputs}
\usage{
\method{predict}{segmentRF}(object, newdata, weights, thresholds = NULL, ...)
}
\arguments{
\item{object}{A segmentRF model}

\item{newdata}{The data to predict}

\item{weights}{A vector or data fram of weights}

\item{thresholds}{A vector of thresholds for each factor}

\item{...}{Additional arguments for predict.randomForest}
}
\value{
A \code{list} of :
\tabular{lll}{
\code{pred} \tab The predicted factor,\cr
\code{prob} \tab A dataframe of probabilities associated with 
\code{weights} \tab The weights associated with each segment \cr
}
}
\description{
Predict a class useing a weighted combination of inputs
}

