% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\title{Segment a data frame based on a numeric vector}
\usage{
segment(x, y, labels = NULL, func = mean)
}
\arguments{
\item{x}{A matrix or data frame with numeric time series in each column}

\item{y}{A numeric vector used to segment columns of \code{x}. Must be
the same length as the columns of \code{x}}

\item{labels}{A dataframe with factor label(s). Labels will be 
be added to each observation in the output dataframe.}

\item{func}{A function to apply to each segment of each column.}
}
\value{
df A dataframe with each corresponds to the supplied function
           applied to the segmented data along with the label
}
\description{
Returns a data frame
}

