% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{segmentRF}
\alias{segmentRF}
\title{Classify segmented time series using a weighted random forest}
\usage{
segmentRF(x, ...)
}
\arguments{
\item{x}{A formula}

\item{...}{Additional arguments for randomForest}
}
\value{
A random forest model of class segmentRF
}
\description{
Classify segmented time series using a weighted random forest
}

